refactor: rename MemberDeleteButton to MemberDeleteMenu
